Ajeitando o código para executar em shell

+++ cadastro_com_registro/src/utilitarias_cadastro_manual/ValidaCPF.java
@@ -27,7 +27,7 @@ public class ValidaCPF {
 			peso = 10;
 
 			for (i = 0; i < 9; i++) {
-				num = (int) (CPF.charAt(i) - 48);
+				num = CPF.charAt(i) - 48;
 				sm = sm + (num * peso);
 				peso = peso - 1;
 			}
@@ -42,7 +42,7 @@ public class ValidaCPF {
 			peso = 11;
 
 			for (i = 0; i < 10; i++) {
-				num = (int) (CPF.charAt(i) - 48);
+				num = CPF.charAt(i) - 48;
 				sm = sm + (num * peso);
 				peso = peso - 1;
 			}
